python annotation

+++ 0x00-python_variable_annotations/102-type_checking.py
@@ -4,7 +4,7 @@
 from typing import List, Tuple
 
 
-def zoom_array(lst: List, factor: int = 2) -> List:
+def zoom_array(lst: List, factor: float = 2.0) -> List:
     """Zooms in on an array by repeating each element 'factor' times."""
     zoomed_in: List = [item for item in lst for _ in range(factor)]
     return zoomed_in
